Redesign jobs index
