+++ app/assets/good_job/modules/application.js
@@ -4,11 +4,13 @@ import renderCharts from "charts";
 import checkboxToggle from "checkbox_toggle";
 import documentReady from "document_ready";
 import showToasts from "toasts";
+import setupPopovers from "popovers";
 import LivePoll from "live_poll";
 
 documentReady(function() {
   renderCharts();
   showToasts();
+  setupPopovers();
   checkboxToggle();
 
   const livePoll = new LivePoll
