+++ app/assets/good_job/modules/popovers.js
@@ -1,0 +1,5 @@
+export default function() {
+  document.querySelectorAll('[data-bs-toggle="popover"]').forEach((el) => {
+    new bootstrap.Popover(el, { html: true })
+  })
+}
